32-bit compat, hostid_read()
Explicitly cast the sizeof in hostid_read() to prevent the
following compiler warning on 32-bit systems.

  module/spl/spl-generic.c:490:10: error: format '%lu' expects
  argument of type 'long unsigned int', but argument 4 has type
  'unsigned int' [-Werror=format]

+++ module/spl/spl-generic.c
@@ -486,8 +486,8 @@ hostid_read(void)
 	if (size < sizeof(HW_HOSTID_MASK)) {
 		printk(KERN_WARNING
 		       "SPL: Ignoring the %s file because it is %llu bytes; "
-		       "expecting %lu bytes instead.\n",
-		       spl_hostid_path, size, sizeof(HW_HOSTID_MASK));
+		       "expecting %lu bytes instead.\n", spl_hostid_path,
+		       size, (unsigned long)sizeof(HW_HOSTID_MASK));
 		kobj_close_file(file);
 		return -3;
 	}
